Emoticon toolbar added, along with many performance improvements.
We've also removed the double linked list from editor nodes, since it is a constant source of frustration, and it's an optimization, and it's too early to optimize. So far it seems to perform fine without it.

We also added an interface/example for how to get the resulting text from the editor only when needed, vs turning the tree into a string with every keystroke.